Se agrega lo siguiente
1. Soporte a integración con captcha para el login.
2. Se actualiza la edición masivas para procesos (scheduler)
3. Se independiza la api para reportes y se implementa (cierre, totalizado)
4. Se actualiza entidad y repository para darle soporte a los reportes.
5. Se agrega el log para seguridad (login, cambiar contraseña, permisos)

+++ masivas-repository/src/main/java/com/novatronic/masivas/backoffice/dto/DetalleConsultaArchivoMasivasDTO.java
@@ -27,8 +27,8 @@ public class DetalleConsultaArchivoMasivasDTO implements Serializable {
     private LocalDateTime fechaProcesada;
     private Long cantidadDeclarado;
     private Long cantidadProcesado;
-    private BigInteger montoProcesado;
-    private BigInteger montoRechazado;
+    private Long montoProcesado;
+    private Long montoRechazado;
     private String estado;
 
 }
